health bar added to core code
